requirements.txt
Django>=5.1.6
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.1
django-cors-headers>=4.3.1
django-extensions>=3.2.3
pyOpenSSL>=24.0.0
Werkzeug>=3.0.1
argon2-cffi>=23.1.0
